+++ examples/multicamera_edex/track_multicamera_r2b.py
@@ -95,7 +95,7 @@ rig.cameras = cameras
 
 odom_cfg = vslam.Odometry.Config(enable_final_landmarks_export = True)
 
-tracker = vslam.Tracker(rig, odom_cfg)
+tracker = vslam.Tracker(rig, vslam.Tracker.Mode.OdometryOnlyRealtime, odom_cfg)
 
 trajectory = []