+++ examples/realsense/run_multicamera.py
@@ -216,7 +216,7 @@ def main() -> None:
         enable_final_landmarks_export=True,
         rectified_stereo_camera=True
     )
-    tracker = vslam.Tracker(rig, odom_cfg)
+    tracker = vslam.Tracker(rig, vslam.Tracker.Mode.OdometryOnlyOffline, odom_cfg)
 
     # Configure all devices
     configure_all_devices(pipelines, configs)